requirements.txt
asn1crypto==0.24.0
boto3==1.7.73
botocore==1.10.73
cffi==1.11.5
cryptography==2.3
docutils==0.14
idna==2.7
jmespath==0.9.3
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-dotenv==0.9.1
s3transfer==0.1.13
six==1.11.0
